move buildAnsibleJobStatus into utils
