+++ package.json
@@ -66,7 +66,7 @@
     "csurf": "^1.11.0",
     "deep-diff": "^1.0.2",
     "del-cli": "^3.0.1",
-    "eslint": "^7.26.0",
+    "eslint": "^7.27.0",
     "eslint-loader": "^4.0.2",
     "express": "^4.17.1",
     "express-handlebars": "^5.3.2",
@@ -121,7 +121,7 @@
     "compression-webpack-plugin": "^7.1.2",
     "copy-webpack-plugin": "^8.1.1",
     "cross-fetch": "^3.1.4",
-    "css-loader": "^5.2.5",
+    "css-loader": "^5.2.6",
     "cypress": "^7.3.0",
     "cypress-fail-fast": "^2.3.2",
     "cypress-multi-reporters": "^1.5.0",
@@ -133,7 +133,7 @@
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-cypress": "^2.11.3",
     "eslint-plugin-html": "^6.1.2",
-    "eslint-plugin-import": "^2.23.2",
+    "eslint-plugin-import": "^2.23.3",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.2.0",
